+++ Desktop-App/LabEyesDesktop.py
@@ -1,32 +1,35 @@
-import pygame
-import button
-from qr_gen import gen, filter
+from src import button
+from src.qr_gen import gen, filter
+from src.videoScanner import scan
 import pygame_textinput as pti
+import pygame
 import cv2
 import sys
-from videoScanner import scan
 
-# create display window
+# height-width	
 SCREEN_HEIGHT = 450
 SCREEN_WIDTH = 800
 TEXTBOX_WIDTH = 272
 TEXTBOX_HEIGHT = 35
 
+# create display window
 screen = pygame.display.set_mode((SCREEN_WIDTH, SCREEN_HEIGHT), pygame.RESIZABLE)
-pygame.display.set_caption('Button Demo')
+pygame.display.set_caption('Lab-Eyes')
 
-backImg = pygame.image.load('back.png').convert_alpha()
+# back button
+backImg = pygame.image.load('assets/textures/back.png').convert_alpha()
 backButton = button.Button(10,10, backImg, 0.07)
 
-# load button images
-genImg = pygame.image.load('QR_GEN.png').convert_alpha()
-scanImg = pygame.image.load('QR_SCAN.png').convert_alpha()
+# load homepage images
+genImg = pygame.image.load('assets/textures/QR_GEN.png').convert_alpha()
+scanImg = pygame.image.load('assets/textures/QR_SCAN.png').convert_alpha()
 
 # text input
 textInputManager = pti.TextInputManager(validator=lambda input: len(input)<=50)
 textinputCustom = pti.TextInputVisualizer(manager=textInputManager,
 										  font_color=(255, 255, 255),
 										  cursor_color=(255, 255, 255))
+
 # game loop
 run = True
 pygame.key.set_repeat(300, 25)
@@ -42,19 +45,22 @@ try:
 			if backButton.draw(screen):
 				page = ''
 
+		# Homepage buttons
 		if page == '':
 			if button.Button(100,200, genImg, 0.25).draw(screen):
-				print('CREATE SCRIPT')
+				print('Generate QR')
 				page = 'gen'
-
 			elif button.Button(500,200, scanImg, 0.25).draw(screen):
-				print('JOIN SCRIPT HERE')
+				print('Scan QR')
 				page = 'scan'
 
+		# QR generator
 		elif page == 'gen':
 			try:
 				currQrImg = pygame.image.load(f'pics/{staticData}.png').convert_alpha()
-				if button.Button(SCREEN_WIDTH/2 - currQrImg.get_width()/2*0.4, SCREEN_HEIGHT*0.65, currQrImg, 0.4).draw(screen):
+				if button.Button(SCREEN_WIDTH/2 - currQrImg.get_width()/2*0.4,
+								 SCREEN_HEIGHT*0.65,
+								 currQrImg, 0.4).draw(screen):
 					print(f'print qr code here')
 			except:
 				pass
@@ -64,7 +70,8 @@ try:
 												  SCREEN_HEIGHT/4,
 												  TEXTBOX_WIDTH,
 												  TEXTBOX_HEIGHT))
-
+		
+		# QR scanner
 		elif page == 'scan':
 			camera = cv2.VideoCapture(0)
 			while True:
@@ -77,7 +84,8 @@ try:
 
 				if button.Button(10,10, backImg, 0.07).draw(screen):
 					page = ''
-					screen = pygame.display.set_mode((SCREEN_WIDTH, SCREEN_HEIGHT), pygame.RESIZABLE)
+					screen = pygame.display.set_mode((SCREEN_WIDTH, SCREEN_HEIGHT),
+													  pygame.RESIZABLE)
 					break
 
 				pygame.display.update()
@@ -86,30 +94,27 @@ try:
 					if event.type == pygame.QUIT:
 						sys.exit(0)
 
-
 		# event handler
 		for event in events:
-			#quit game
+			# quit game
 			if event.type == pygame.QUIT:
 				run = False
 
+			# enter data
 			if event.type == pygame.KEYDOWN and event.key == pygame.K_RETURN:
 				if page == 'gen':
 					data = filter(textinputCustom.value)
 					staticData = data
-					print(f"Name = {data}")
+					print(f"Data = {data}")
 					gen(data)
 					textinputCustom.value = ''
 
 			if event.type == pygame.VIDEORESIZE:
-				old_surface_saved = screen
 				screen = pygame.display.set_mode((event.w, event.h),
-												pygame.RESIZABLE)
-				screen.blit(old_surface_saved, (0,0))
-				del old_surface_saved
+												  pygame.RESIZABLE)
 
 		pygame.display.update()
-		
+
 except (KeyboardInterrupt, SystemExit):
     pygame.quit()
     cv2.destroyAllWindows()
