+++ Desktop-App/assets/textures/QR_GEN.png
