+++ Desktop-App/assets/textures/QR_SCAN.png
