--- Desktop-App/config.py
@@ -1,2 +1,0 @@
-newData = ''
-oldData = ''
