+++ Desktop-App/src/config.py
@@ -1,0 +1,6 @@
+#------------------
+# Global variables 
+#------------------
+
+newData = ''
+oldData = ''
