+++ Desktop-App/src/qr_gen.py
