+++ Desktop-App/src/scanner_frame.py
